tun-device: Improve handling of IPv6 addresses
struct ifreq can't be used for IPv6 as the ifr_addr member is not large
enough.  Actually, configuring an IPv6 via an AF_INET socket won't work
anyway.  And unfortunately it is not standardized how IPv6 addresses are
installed, so we have to do this pretty differently on Linux and on BSD.