WIP: cng: DH implementation
Because CNG does not allow us to export the raw DH secret, an implementation
using our DH interface seems unfeasible. Changing the DH interface to do
key derivation could work, but is non-trivial.

As cng without DH does not allow us to remove the gmp/openssl/... dependency,
this is probably a dead-end for cng.