settings: Ignore current text in error handling after EOF reached
When an error occurs after we reached the end of file the pointer
yyget_text() returns is invalid as that memory has already been freed.
Unfortunately, yyget_leng() still returns 1, although the text actually
only points to '\0', so not exactly useful.  Therefore, instead of
completely changing the memory management in the parser, we use a heuristic
to detect this situation and then ignore the text.