WIP
Without this the xauth-eap plugin, or rather the instantiated EAP method is
not able to access the CA certificates in other credential sets as only an
exclusive local set is loaded (in order to not affect the eap method
which runs in the same thread).

Another problem is that the server identity is not stored in the
auth_cfg object that results from the XAuth exchange (add_auth_cfg()
function in xauth.c), so the aaa_identity constraint check will
eventually fail.