time: Use CLOCK_BOOTTIME if available
This clock works like CLOCK_MONOTONIC (i.e. is not affected by time
adjustments) but includes times when the system was suspended.  So
we can schedule and measure the delay between events more accurately
e.g. on mobile devices like Android that are often suspended.

We currently don't check for this via configure script.